worker/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.121.2
h11==0.16.0
httptools==0.7.1
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
schedule
